+++ CHANGELOG.md
@@ -125,3 +125,4 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fixes incorrect stock deallocation when multiple order lines share the same ProductVariant - #17657 by @korycins
 - Decrease allocations for lines with inventory tracking disabled, if allocations exist - #17657 by @korycins
 - Make token generator class configurable - #17701 by @wcislo-saleor
+- Add a new `variant` field on `AssignedVariantAttributeValue`. First part of a simplification of Attribute - ProductVariant relation from #12881. by @aniav